Python_tutorial/Assignments/Machine_learning/requirements.txt
aiofiles==22.1.0
anyio==3.6.2
asgiref==3.6.0
asttokens==2.2.1
backcall==0.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
ci-info==0.3.0
click==8.1.3
comm==0.1.2
configobj==5.0.7
configparser==5.3.0
debugpy==1.6.5
decorator==5.1.1
Django==4.1.5
entrypoints==0.4
etelemetry==0.3.0
executing==1.2.0
filelock==3.9.0
frontend==0.0.3
fst-pso==1.8.1
future==0.18.3
FuzzyTM==2.0.5
gensim==4.3.0
h11==0.14.0
httplib2==0.21.0
idna==3.4
ipykernel==6.20.2
ipython==8.8.0
isodate==0.6.1
itsdangerous==2.1.2
jedi==0.18.2
joblib==1.2.0
jupyter_client==7.4.9
jupyter_core==5.1.3
looseversion==1.0.3
lxml==4.9.2
matplotlib-inline==0.1.6
miniful==0.0.6
nest-asyncio==1.5.6
networkx==3.0
nibabel==5.0.0
nipype==1.8.5
nltk==3.8.1
numpy==1.24.1
packaging==23.0
pandas==1.5.2
parso==0.8.3
pathlib==1.0.1
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.6.2
prompt-toolkit==3.0.36
prov==2.0.0
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pydot==1.4.2
pyFUME==0.2.25
Pygments==2.14.0
PyMuPDF==1.21.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2022.7.1
pyxnat==1.5
pyzmq==25.0.0
rdflib==6.2.0
regex==2022.10.31
requests==2.28.2
scikit-learn==1.2.0
scipy==1.10.0
simpful==2.9.0
simplejson==3.18.1
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
sqlparse==0.4.3
stack-data==0.6.2
starlette==0.23.1
threadpoolctl==3.1.0
tika==2.6.0
tornado==6.2
tqdm==4.64.1
traitlets==5.8.1
traits==6.3.2
urllib3==1.26.14
uvicorn==0.20.0
wcwidth==0.2.6
